requirements.txt
openai-whisper==20250625
scikit-learn>=1.7.0
numpy>=1.22.0
scipy>=1.8.0
joblib>=1.2.0
threadpoolctl>=3.1.0
